backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
textstat
nltk
groq
sentence-transformers